+++ CHANGELOG.md
@@ -12,5 +12,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Unable to export when name contains invalid path characters
+- Remove overrides folder from exports when it's empty
 
 ### Misc